[backend] Introduce master lock
